[clean] Abstractions/interface allocated to several modules
