+++ speculos/mcu/display.py
@@ -1,22 +1,28 @@
+from __future__ import annotations
+
 import io
+from abc import ABC, abstractmethod
 from PIL import Image
-from typing import Dict, NamedTuple, Optional, Tuple
+from socket import socket
+from typing import Any, Dict, IO, List, Optional, Tuple, Union
+
+from .struct import DisplayArgs, MODELS, ServerArgs, TextEvent
 
 
-class Model(NamedTuple):
-    name: str
-    screen_size: Tuple[int, int]
-    box_position: Tuple[int, int]
-    box_size: Tuple[int, int]
+class IODevice(ABC):
 
+    @property
+    @abstractmethod
+    def file(self) -> Union[IO[bytes], socket]:
+        raise NotImplementedError()
 
-MODELS: Dict[str, Model] = {
-    'nanos': Model('Nano S', (128, 32), (20, 13), (100, 26)),
-    'nanox': Model('Nano X', (128, 64), (5, 5), (10, 10)),
-    'nanosp': Model('Nano SP', (128, 64), (5, 5), (10, 10)),
-    'blue': Model('Blue', (320, 480), (13, 13), (26, 26)),
-    'stax': Model('Stax', (400, 672), (13, 13), (26, 26)),
-}
+    @property
+    def fileno(self) -> int:
+        return self.file.fileno()
+
+    @abstractmethod
+    def can_read(self, s: int, screen: Display) -> None:
+        pass
 
 
 COLORS: Dict[str, int] = {
@@ -136,3 +142,98 @@ class FrameBuffer:
                w: Optional[int] = None,
                h: Optional[int] = None) -> bool:
         raise NotImplementedError()
+
+
+class GraphicLibrary(ABC):
+
+    def __init__(self, fb: FrameBuffer, size: Tuple[int, int], model: str):
+        self._fb = fb
+        self.SCREEN_WIDTH, self.SCREEN_HEIGHT = size
+        self.model = model
+
+    @property
+    def fb(self) -> FrameBuffer:
+        return self._fb
+
+    @abstractmethod
+    def refresh(self, data: bytes) -> bool:
+        pass
+
+    def update_screenshot(self) -> None:
+        self.fb.update_screenshot()
+
+    def update_public_screenshot(self) -> None:
+        self.fb.update_public_screenshot()
+
+    def take_screenshot(self) -> Tuple[Tuple[int, int], bytes]:
+        return self.fb.take_screenshot()
+
+
+class Display(ABC):
+    def __init__(self, display: DisplayArgs, server: ServerArgs) -> None:
+        self.notifiers: Dict[int, Any] = {}
+        self._server = server
+        self._display = display
+
+    @property
+    def apdu(self) -> Any:  # ApduServer:
+        return self._server.apdu
+
+    @property
+    def seph(self) -> Any:  # SeProxyHal:
+        return self._server.seph
+
+    @property
+    def model(self) -> str:
+        return self._display.model
+
+    @property
+    def force_full_ocr(self) -> bool:
+        return self._display.force_full_ocr
+
+    @property
+    def disable_tesseract(self) -> bool:
+        return self._display.disable_tesseract
+
+    @property
+    def rendering(self):
+        return self._display.rendering
+
+    @property
+    @abstractmethod
+    def gl(self) -> GraphicLibrary:
+        pass
+
+    @abstractmethod
+    def display_status(self, data: bytes) -> List[TextEvent]:
+        raise NotImplementedError()
+
+    @abstractmethod
+    def display_raw_status(self, data: bytes):
+        pass
+
+    @abstractmethod
+    def screen_update(self) -> bool:
+        pass
+
+    @abstractmethod
+    def run(self) -> None:
+        pass
+
+    def add_notifier(self, klass: IODevice):
+        assert klass.fileno not in self.notifiers
+        self.notifiers[klass.fileno] = klass
+
+    def remove_notifier(self, fd: int):
+        self.notifiers.pop(fd)
+
+    def _init_notifiers(self, args: ServerArgs) -> None:
+        for klass in args._asdict().values():
+            if klass:
+                self.add_notifier(klass)
+
+    def forward_to_app(self, packet: bytes):
+        self.seph.to_app(packet)
+
+    def forward_to_apdu_client(self, packet):
+        self.apdu.forward_to_client(packet)
